Big cleanup; Removed some base classes
